+++ src/project/src/repository/AccountRepo.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 public interface AccountRepo {
     Optional<Account> accessAccount(int id);
 
-    int queryBalance(int id);
+    double queryBalance(int id);
 }
