+++ src/project/src/service/TransferManager.java
@@ -1,5 +1,5 @@
 package src.service;
 
 public interface TransferManager {
-    void transfer(int fromId, int toId, int amount);
+    void transfer(int fromId, int toId, double amount);
 }
